barneystinson/requirements/base.txt
boto==2.45.0
boto3==1.4.3
Django==1.10.5
django-model-utils==2.6.1
django-storages==1.5.2
djangorestframework==3.5.3
Pillow==4.0.0
psycopg2==2.6.2
pytz==2016.10
Unipath==1.1
